% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dictionary_size}
\alias{dictionary_size}
\title{Helper function to get the size of compressed paths including dictionary}
\usage{
dictionary_size(x, maxdepth = 1, compressed = NULL)
}
\description{
Helper function to get the size of compressed paths including dictionary
}
\keyword{internal}
