% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{named_vector_to_data_frame}
\alias{named_vector_to_data_frame}
\title{Helper function to convert a named vector to a data frame}
\usage{
named_vector_to_data_frame(x)
}
\description{
Helper function to convert a named vector to a data frame
}
\keyword{internal}
