% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{order_by_score}
\alias{order_by_score}
\title{Order Data Frame Decreasingly by Score Column}
\usage{
order_by_score(df, column = "score2")
}
\arguments{
\item{df}{data frame}

\item{column}{name of column by which to order decreasingly. Default:
"score2"}
}
\description{
Order Data Frame Decreasingly by Score Column
}
