% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{order_decreasingly_by}
\alias{order_decreasingly_by}
\title{Order Data Frame Decreasingly by one Column}
\usage{
order_decreasingly_by(df, column)
}
\arguments{
\item{df}{data frame}

\item{column}{name of column by which to order decreasingly.}
}
\description{
Order Data Frame Decreasingly by one Column
}
