% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_compressed}
\alias{print_compressed}
\title{Helper function to print compressed paths and dictionary}
\usage{
print_compressed(x)
}
\description{
Helper function to print compressed paths and dictionary
}
\keyword{internal}
