% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_dict}
\alias{print_dict}
\title{Helper function to print the dictionary to the console}
\usage{
print_dict(x, maxdepth = 1, compressed = NULL)
}
\description{
Helper function to print the dictionary to the console
}
\keyword{internal}
