% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_paths.R
\name{random_paths}
\alias{random_paths}
\title{Create Random File Paths Using English Words}
\usage{
random_paths(max_depth = 5, min_chars = 5, max_elements = 10,
  depth_to_leaf_weight = function(depth) 1.2^depth)
}
\arguments{
\item{max_depth}{maximum path depth}

\item{min_chars}{least number of characters per folder or file name}

\item{max_elements}{maximum number of elements (files or subfolders) in a
folder}

\item{depth_to_leaf_weight}{function that calculates a weight from the given
path depth. The weight is used to increase the probability of a folder
element to be a file and not a subdirectory. By default the weight is
calculated by 1.2^depth, i.e. for a folder in depth 10 it is about six
times (1.2^10 = 6.19) more probable of its elements to be files instead of
subfolders}
}
\description{
Create Random File Paths Using English Words
}
\examples{
paths <- kwb.pathdict::random_paths(max_depth = 5)
table(lengths(kwb.file::split_paths(paths)))
}
