% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary_one_by_one.R
\name{rescore_and_reorder_frequency_data}
\alias{rescore_and_reorder_frequency_data}
\title{Rescore and Reorder Frequency Data}
\usage{
rescore_and_reorder_frequency_data(frequency_data, placeholder_size)
}
\arguments{
\item{frequency_data}{data frame with columns \code{length} and \code{count}}

\item{placeholder_size}{size of placeholder in number of characters. The path
length will be reduced by this value before being multiplied with the count
to calculate the score.}
}
\description{
Rescore and Reorder Frequency Data
}
