% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_path.R
\name{starts_with_parts}
\alias{starts_with_parts}
\title{Do Subfolder List Elements Start with Given Folder Names?}
\usage{
starts_with_parts(parts, elements)
}
\arguments{
\item{parts}{list of list of character as returned by
\code{\link[kwb.file]{split_paths}} or a matrix of character representing
the subfolder names at the different folder depths as returned by
\code{\link[kwb.file]{to_subdir_matrix}}.}

\item{elements}{vector of character giving the sequence of strings to be
found in \code{parts}}
}
\value{
vector of logical as long as \code{parts} containing \code{TRUE} at
  positions \code{i} for which \code{all(parts[[i]][seq_along(elements)] ==
  elements)} is \code{TRUE}
}
\description{
Do Subfolder List Elements Start with Given Folder Names?
}
\examples{
parts <- strsplit(c("a/b/c", "a/b/d", "b/c"), "/")
starts_with_parts(parts, elements = c("a", "b"))
starts_with_parts(parts, elements = c("b", "c"))

subdir_matrix <- kwb.file::to_subdir_matrix(parts)
starts_with_parts(subdir_matrix, elements = c("a", "b"))
starts_with_parts(subdir_matrix, elements = c("b", "c"))

}
