% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_paths.R
\name{to_dictionary}
\alias{to_dictionary}
\title{Create Dictionary from Unique Strings}
\usage{
to_dictionary(x, prefix = "a", leading_zeros = FALSE)
}
\arguments{
\item{x}{vector of strings}

\item{prefix}{prefix to be given to the keys in the dictionary.
Default: "a"}

\item{leading_zeros}{whether to make all keys in the dictionary have same
length by adding leading zeros to the keys. Default: \code{FALSE}}
}
\description{
Create Dictionary from Unique Strings
}
\examples{
# Define input strings
x <- c("elephant", "mouse", "cat", "cat", "cat", "mouse", "cat", "cat")

# Create a dictionary for the unique values in x
kwb.pathdict:::to_dictionary(x)

# Note that "cat" is the first entry because it has the highest "importance"
kwb.pathdict:::sorted_importance(x)

}
