% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dictionary.R
\name{use_dictionary}
\alias{use_dictionary}
\title{Substitute Values that are in a Dictionary with their Keys}
\usage{
use_dictionary(x, dict, method = "full")
}
\arguments{
\item{x}{vector of character}

\item{dict}{list of key = value pairs. Values of this list that are found in
\code{x} are replaced by the placeholder "<key>" in \code{x}.}

\item{method}{method to be applied, must be one of "full" or "part".
If "full", the full values must match, otherwise the values in \code{dict}
are interpreted as patterns that are matched against the values in \code{x}
and the matching parts are replaced with the corresponding "<key>"
placeholder}
}
\value{
\code{x} in which values or parts of the values are replaced with
  their short forms as they are defined in the dictionary \code{dict}
}
\description{
Substitute Values that are in a Dictionary with their Keys
}
\examples{
# Define a vector of long values
x <- c("What a nice day", "Have a nice day", "Good morning")

# Define short forms for full or partial values
dict_full <- list(wand = "What a nice day", gm = "Good morning")
dict_part <- list(w = "What", nd = "nice day", g = "Good")

# Replace long form values with their short forms
kwb.pathdict:::use_dictionary(x, dict_full, method = "full")
kwb.pathdict:::use_dictionary(x, dict_part, method = "part")
}
